Test of Parameters in Division
Prelim. conclusions:
When epsilon and denominator approach each other, the absolute error increases drastically.
When the denominator is small and the numerator is large, the absolute error increases.
To be considered: When does the denominator become so small? Do the weights have any influence wrt. this? E.g. for division, can W become [1, -x], x->0 ?